+++ relearn/source/sim/SubdomainFromNeuronDensity.cpp
@@ -131,13 +131,13 @@ void SubdomainFromNeuronDensity::place_neurons_in_area(
     }
 
     const auto current_num_neurons = get_number_placed_neurons();
-    const auto former_ex_neurons = current_num_neurons * get_ratio_placed_excitatory_neurons();
+    const auto former_ex_neurons = static_cast<double>(current_num_neurons) * get_ratio_placed_excitatory_neurons();
 
     const auto new_num_neurons = current_num_neurons + placed_neurons;
 
     set_number_placed_neurons(new_num_neurons);
 
-    const auto now_ex_neurons = former_ex_neurons + expected_number_ex;
+    const auto now_ex_neurons = former_ex_neurons + static_cast<double>(expected_number_ex);
     // const auto now_in_neurons = former_in_neurons + placed_in_neurons;
 
     const auto current_frac_ex = static_cast<double>(now_ex_neurons) / static_cast<double>(new_num_neurons);
